# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MS365 Access API",
    "description": "API for accessing Microsoft 365 services (Email, Calendar, OneDrive)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/mail/folders": {
      "get": {
        "tags": [
          "mail"
        ],
        "summary": "List Folders",
        "operationId": "list_folders_mail_folders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/mail/folders/resolve/{name}": {
      "get": {
        "tags": [
          "mail"
        ],
        "summary": "Resolve Folder Name",
        "description": "Resolve a well-known folder name to its folder object including ID.\n\nWell-known folder names include: inbox, drafts, sentitems, deleteditems,\njunkemail, archive, outbox, etc.\n\nThis is useful because MS Graph API can cache results when querying by\nwell-known name. Using the actual folder ID ensures fresh results.",
        "operationId": "resolve_folder_name_mail_folders_resolve__name__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mail/messages": {
      "get": {
        "tags": [
          "mail"
        ],
        "summary": "List Messages",
        "description": "List messages from a folder.\n\nNote: MS Graph API can cache results when querying folders by well-known name.\nThis endpoint resolves folder names to IDs internally to ensure fresh results.",
        "operationId": "list_messages_mail_messages_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "folder",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Well-known folder name (inbox, archive, junkemail, etc.). Will be resolved to folder_id internally to avoid MS Graph caching issues.",
              "title": "Folder"
            },
            "description": "Well-known folder name (inbox, archive, junkemail, etc.). Will be resolved to folder_id internally to avoid MS Graph caching issues."
          },
          {
            "name": "folder_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Actual folder ID. Takes precedence over 'folder' if both provided.",
              "title": "Folder Id"
            },
            "description": "Actual folder ID. Takes precedence over 'folder' if both provided."
          },
          {
            "name": "top",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 25,
              "title": "Top"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Filter"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "receivedDateTime desc",
              "title": "Order By"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "mail"
        ],
        "summary": "Send Mail",
        "operationId": "send_mail_mail_messages_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mail/messages/{message_id}": {
      "get": {
        "tags": [
          "mail"
        ],
        "summary": "Get Message",
        "operationId": "get_message_mail_messages__message_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "mail"
        ],
        "summary": "Update Message",
        "operationId": "update_message_mail_messages__message_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "mail"
        ],
        "summary": "Delete Message",
        "operationId": "delete_message_mail_messages__message_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mail/drafts": {
      "post": {
        "tags": [
          "mail"
        ],
        "summary": "Create Draft",
        "description": "Create a new draft message in Drafts folder (does not send).\n\nRequires write:draft permission. To send the draft later, use\nPOST /mail/messages/{draft_id}/send (requires write:mail).",
        "operationId": "create_draft_mail_drafts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/mail/messages/{message_id}/send": {
      "post": {
        "tags": [
          "mail"
        ],
        "summary": "Send Draft",
        "description": "Send an existing draft message.",
        "operationId": "send_draft_mail_messages__message_id__send_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mail/messages/{message_id}/draftReply": {
      "post": {
        "tags": [
          "mail"
        ],
        "summary": "Create Reply Draft",
        "description": "Create a draft reply to a message (does not send it).\n\nReturns the draft message object with its ID. Use PATCH /mail/messages/{draft_id}\nto update the body before sending.",
        "operationId": "create_reply_draft_mail_messages__message_id__draftReply_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          },
          {
            "name": "reply_all",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Create reply-all draft instead of reply",
              "default": false,
              "title": "Reply All"
            },
            "description": "Create reply-all draft instead of reply"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mail/messages/{message_id}/reply": {
      "post": {
        "tags": [
          "mail"
        ],
        "summary": "Reply To Message",
        "operationId": "reply_to_message_mail_messages__message_id__reply_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyMailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mail/messages/{message_id}/forward": {
      "post": {
        "tags": [
          "mail"
        ],
        "summary": "Forward Message",
        "operationId": "forward_message_mail_messages__message_id__forward_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForwardMailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mail/messages/{message_id}/move": {
      "post": {
        "tags": [
          "mail"
        ],
        "summary": "Move Message",
        "description": "Move a message to a destination folder.\n\nThe destination_folder_id can be either an actual folder ID or a well-known\nfolder name (inbox, archive, junkemail, etc.).\n\nIf verify=True (default), the response will include a 'verified' field\nindicating whether the message's parentFolderId matches the destination.",
        "operationId": "move_message_mail_messages__message_id__move_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          },
          {
            "name": "verify",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Verify the move by checking parentFolderId after",
              "default": true,
              "title": "Verify"
            },
            "description": "Verify the move by checking parentFolderId after"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveMailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MoveMailRequest": {
        "properties": {
          "destination_folder_id": {
            "type": "string",
            "title": "Destination Folder Id"
          }
        },
        "type": "object",
        "required": [
          "destination_folder_id"
        ],
        "title": "MoveMailRequest"
      },
...
      "ReplyMailRequest": {
        "properties": {
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "reply_all": {
            "type": "boolean",
            "title": "Reply All",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "comment"
        ],
        "title": "ReplyMailRequest"
      },
...
      "SendMailRequest": {
        "properties": {
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "body_type": {
            "type": "string",
            "title": "Body Type",
            "default": "HTML"
          },
          "to_recipients": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "To Recipients"
          },
          "cc_recipients": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cc Recipients",
            "default": []
          },
          "bcc_recipients": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Bcc Recipients",
            "default": []
          },
          "importance": {
            "type": "string",
            "title": "Importance",
            "default": "normal"
          },
          "save_to_sent_items": {
            "type": "boolean",
            "title": "Save To Sent Items",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "subject",
          "body",
          "to_recipients"
        ],
        "title": "SendMailRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}